+++ server/app/src/modules/statements/firstRun/initiateDatabase/create_schema.js
