+++ server/app/src/modules/statements/firstRun/initiateDatabase/create_table_fannyPackz.js
