+++ server/app/src/modules/statements/firstRun/initiateDatabase/drop_database.js
