+++ server/app/src/modules/statements/firstRun/initiateDatabase/kill_connection.js
