+++ server/app/src/router/fannyPack/addFannyPack/utli/create_schema_user_fannyPack.js
@@ -11,7 +11,7 @@
  * │   │   Table - account_type - account_type_id
  * │   │   Table - account_record - account_id
  * │   │   Table - account_One - account_id
- * 
+ *  
     Create - FannyPack
     - Requirement
         - > fannyPackName, userSerialID, fannyPackSerial
@@ -30,18 +30,6 @@
 // Create Schema - create_schema_fannyPack
 // Function - Insert user FannyPack to FannyPack record
 const create_schema_fannyPack = function(fannyPackSerial) {
-  let pushD = {}
-  let fannyPack = "FannyPack-" + fannyPackSerial;
-  const create_Schema = "CREATE SCHEMA IF NOT EXISTS " + fannyPack + " AUTHORIZATION " + app_db.blingBlaw.user + ";";
-  // blaze
-  danzillaDB.pool.query(create_Schema,
-      function (err, Results) {
-        if (!err && Results) { // If no errors and Results == Good
-          pushD = { checked: "\n checked", results: Results }
-        } else if (err) { // if any errors
-          pushD = { checked: "\n ", results: err }
-        }
-        console.log(JSON.stringify(pushD));
-  });
+  console.log("LOOOOL: " + fannyPackSerial);
 }
 module.exports = create_schema_fannyPack;
