+++ server/app/src/router/firstrun/initiateDatabase/index.js
@@ -29,16 +29,14 @@
 */
 // Async Waterfall
 const async = require('async');
-// Import FirstRun
-const FirstRun = {
-    kill_connection: require("./utli/kill_connection"),
-    drop_Database: require("./utli/drop_database"),
-    create_Database: require("./utli/create_database"),
-    create_Schema: require("./utli/create_schema"),
-    create_Table_UserAuth: require("./utli/create_table_userAuth"),
-    create_Table_UserDetails: require("./utli/create_table_userDetails"),
-    create_Table_fannyPackz: require("./utli/create_table_fannyPackz")
-}
+// bling #build
+kill_connection = require("../../../modules/statements/firstRun/initiateDatabase/kill_connection");
+drop_Database = require("../../../modules/statements/firstRun/initiateDatabase/drop_database");
+create_Database = require("../../../modules/statements/firstRun/initiateDatabase/create_database");
+create_Schema = require("../../../modules/statements/firstRun/initiateDatabase/create_schema");
+create_Table_UserAuth = require("../../../modules/statements/firstRun/initiateDatabase/create_table_userAuth");
+create_Table_UserDetails = require("../../../modules/statements/firstRun/initiateDatabase/create_table_userDetails");
+create_Table_fannyPackz = require("../../../modules/statements/firstRun/initiateDatabase/create_table_fannyPackz");
 // Collect FirstRunCheck
 const First_Run_Check_Result = []
 // Initiate Database 
@@ -47,25 +45,26 @@ const initiateDB = function (req, res, next) {
     async.waterfall([
             // Drop Database
         function (callback) {
-            FirstRun.drop_Database(callback, First_Run_Check_Result)
+            drop_Database(callback, First_Run_Check_Result)
         },  // Create Database
         function (result_drop_Database, callback) {
             // If any active session - Kill
             if (result_drop_Database[0].results.code == "55006"){
-                FirstRun.kill_connection(First_Run_Check_Result);
-            } FirstRun.create_Database(callback, First_Run_Check_Result);
+                kill_connection(First_Run_Check_Result);
+            } 
+            create_Database(callback, First_Run_Check_Result);
         },  // Create Schema - create_Schema
         function (result_create_Database, callback) {
-            FirstRun.create_Schema(callback, First_Run_Check_Result);
+            create_Schema(callback, First_Run_Check_Result);
         }, // Create Table - create_Table_UserAuth
         function (result_create_Schema, callback) {
-            FirstRun.create_Table_UserAuth(callback, First_Run_Check_Result);
+            create_Table_UserAuth(callback, First_Run_Check_Result);
         }, // Create Table - create_Table_UserDetails
         function (result_create_Table_UserAuth, callback) {
-            FirstRun.create_Table_UserDetails(callback, First_Run_Check_Result);
+            create_Table_UserDetails(callback, First_Run_Check_Result);
         }, // Create Table - create_Table_fannyPackz
         function (result_create_Table_UserDetails, callback) {
-            FirstRun.create_Table_fannyPackz(callback, First_Run_Check_Result);
+            create_Table_fannyPackz(callback, First_Run_Check_Result);
         }
     ], function (err, Results) {
         let pageMesage = "lalalala~ Initial DB";
